Changed the type of the deleted field to bigint

+++ Pinnacle/src/main/java/com/cfive/pinnacle/entity/Attendance.java
@@ -71,7 +71,7 @@ public class Attendance implements Serializable {
 
     @TableField("deleted")
     @TableLogic(value = "0",delval = "1")
-    private Integer deleted;
+    private Long deleted;
 
     @TableField("version")
     @Version
